Revert "This makes everything much more dynamic in terms of how the correct JSON schema is determined."
This reverts commit a8cab003d9fbf159b78ee5e8b58d9632a3ad203d.
